TIKA-1532: Moving DIFParser ContentHandler and Test to dif package
