+++ tika-parsers/src/main/java/org/apache/tika/parser/dif/DIFContentHandler.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.tika.sax;
+package org.apache.tika.parser.dif;
 
 import java.util.Stack;
