+++ tika-parsers/src/main/resources/META-INF/services/org.apache.tika.parser.Parser
@@ -52,7 +52,7 @@ org.apache.tika.parser.rtf.RTFParser
 org.apache.tika.parser.txt.TXTParser
 org.apache.tika.parser.video.FLVParser
 org.apache.tika.parser.xml.DcXMLParser
-org.apache.tika.parser.xml.DIFParser
+org.apache.tika.parser.dif.DIFParser
 org.apache.tika.parser.xml.FictionBookParser
 org.apache.tika.parser.chm.ChmParser
 org.apache.tika.parser.code.SourceCodeParser
