+++ tika-parsers/src/test/java/org/apache/tika/parser/dif/DIFParserTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.tika.parser.xml;
+package org.apache.tika.parser.dif;
 
 import static org.junit.Assert.assertEquals;
